separate CostRegistry
